lib-schema) complete rewrite of the lib
This is a complete reimplemented version of the lib-schema.
As following the json schema spec are not applicable for the use
case and are not compatible with the Rust type system because of the
dynamic typing of json itself.

This apporoach uses a custom validator, and registry. It also include a
ValueStore for storing all the value into a object like structure, and
also the metadata needed for the validation.

It is similiar to what serde_json have implementend (in fact some of the
function relies on the serde_json API) but with different and more
personal API apporoach but it is very similiar.